LayoutTests/fast/images/image-extraction/mac/image-overlay-text-disables-app-highlight-menu-items-expected.txt

PASS foundTitle1 is true
PASS foundTitle2 is true
PASS Selected image overlay
PASS foundTitle1 is false
PASS foundTitle2 is false
PASS successfullyParsed is true

TEST COMPLETE
